test: improve parser tests
